player Movement completion  with stamina bar
Successfully implemented player movement with an advanced first-person view camera, which includes the integration of a stamina bar. This stamina system depletes when the player runs excessively, leading to a gradual reduction in movement speed. This addition enhances the overall realism of the gameplay experience.

Utilized character controller technology to leverage its precision control and collision handling capabilities. Additionally, harnessed the performance optimization advantages it offers.